docs/research/baml/2025-12-02-multimodal-evals/pyproject.toml
[project]
name = "2025-12-02-multimodal-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baml-py>=0.212.0",
    "datasets>=4.4.0",
    "kagglehub>=0.3.13",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.4.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "ruff>=0.14.3",
]
